Ensure consistent naming
I switched back and forth between "literatevim" and "literatevimrc"
during development of this plugin. I felt "literatevimrc" was more
accurate in the end. I, evidently, missed a few renames which could
cause trouble for end users.

Bumping version to 0.0.2 because, although this is technically
incompatible with version 0.0.1, version 0.0.1 was broken and this does
not change a working API.

Fixes #1