Remove erroneous install instructions
The `runtime` command expects a filename which it will try to execute
whereas `execute` calls an existing function. Since `literatevimrc.vim`
should be in the autoload directory, calling the `load` function
directly is the correct invocation in a `~/.vimrc`.

This removes an erroneous install instruction in the README. Thanks to
@NoahTheDuke for spotting my sloppy errors :)

Addresses #1